refactor(storage): improve config checkers
- Add missing textdomain
