+++ service/lib/agama/storage/config_checkers/drive.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Copyright (c) [2024] SUSE LLC
+# Copyright (c) [2024-2025] SUSE LLC
 #
 # All Rights Reserved.
 #
@@ -36,12 +36,12 @@ module Agama
         include WithSearch
 
         # @param config [Configs::Drive]
-        # @param storage_config [Storage::Config]
         # @param product_config [Agama::Config]
-        def initialize(config, storage_config, product_config)
-          super(storage_config, product_config)
+        def initialize(config, product_config)
+          super()
 
           @config = config
+          @product_config = product_config
         end
 
         # Drive config issues.
@@ -60,6 +60,9 @@ module Agama
 
         # @return [Configs::Drive]
         attr_reader :config
+
+        # @return [Agama::Config]
+        attr_reader :product_config
       end
     end
   end
