+++ service/lib/agama/storage/config_checkers/partition.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Copyright (c) [2024] SUSE LLC
+# Copyright (c) [2024-2025] SUSE LLC
 #
 # All Rights Reserved.
 #
@@ -34,12 +34,12 @@ module Agama
         include WithSearch
 
         # @param config [Configs::Partition]
-        # @param storage_config [Storage::Config]
         # @param product_config [Agama::Config]
-        def initialize(config, storage_config, product_config)
-          super(storage_config, product_config)
+        def initialize(config, product_config)
+          super()
 
           @config = config
+          @product_config = product_config
         end
 
         # Partition config issues.
@@ -57,6 +57,9 @@ module Agama
 
         # @return [Configs::Partition]
         attr_reader :config
+
+        # @return [Agama::Config]
+        attr_reader :product_config
       end
     end
   end
