+++ service/lib/agama/storage/config_checkers/with_encryption.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Copyright (c) [2024] SUSE LLC
+# Copyright (c) [2024-2025] SUSE LLC
 #
 # All Rights Reserved.
 #
@@ -19,7 +19,7 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require "agama/storage/config_checkers/filesystem"
+require "agama/storage/config_checkers/encryption"
 
 module Agama
   module Storage
@@ -28,7 +28,7 @@ module Agama
       module WithEncryption
         # @return [Array<Issue>]
         def encryption_issues
-          ConfigCheckers::Encryption.new(config, storage_config, product_config).issues
+          ConfigCheckers::Encryption.new(config).issues
         end
       end
     end
