+++ service/lib/agama/storage/config_checkers/with_filesystem.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Copyright (c) [2024] SUSE LLC
+# Copyright (c) [2024-2025] SUSE LLC
 #
 # All Rights Reserved.
 #
@@ -28,7 +28,7 @@ module Agama
       module WithFilesystem
         # @return [Array<Issue>]
         def filesystem_issues
-          ConfigCheckers::Filesystem.new(config, storage_config, product_config).issues
+          ConfigCheckers::Filesystem.new(config, product_config).issues
         end
       end
     end
